queue_server/requirements.txt
aio-pika
aiormq
pydantic
python-dotenv
redis
